Rename ApplyRoute to apply.route
